Update module_troops.py
Fixes the "Random End" lord bug.

Why this happened: At the game start random lords get their alive slot set to 0. This is to disable them in the game start. The problem is, that the last placeholder troop of the list (Random End) didn't had that set up, and since he was from fac_kingdom_1 as especified in the troops file, he spawned when the script that assigns fiefs to lords was called.